template/{{cookiecutter.project_slug}}/requirements.txt
Brotli==1.0.9
click==7.1.2
dash==1.16.0
dash-bootstrap-components==0.10.6
dash-core-components==1.12.0
dash-html-components==1.1.1
dash-renderer==1.8.0
dash-table==4.10.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
pandas==1.1.2
plotly==4.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
retrying==1.3.3
six==1.15.0
Werkzeug==1.0.1
